Update prompts for GPT-5 tests
Contrary to previous models, GPT-5 often tries to return both the move
action and FOUND action at the same time (with reasoning in form of I'll
move x to and then return found). Added a line to the prompt, that
underlines the fact, that only one action can be returned at time.